Update tables-feature-support.md
Currently, AMA is incompatible with data transformation. This should be fixed at some point in June, but we need to fix the documentation. Here it is correct: https://docs.microsoft.com/en-us/azure/azure-monitor/logs/ingestion-time-transformations#supported-workflows  - "Ingestion-time transformation is applied to any workflow that doesn't currently use a data collection rule to send data to a supported table. Any transformation on a workspace will be ignored for these workflows."